AND-6588 Reuse network providers components
