+++ blockchain/src/androidTest/kotlin/com/tangem/blockchain/blockchains/cardano/CardanoTransactionTest.kt
@@ -1,6 +1,6 @@
 package com.tangem.blockchain.blockchains.cardano
 
-import com.tangem.blockchain.blockchains.cardano.network.CardanoUnspentOutput
+import com.tangem.blockchain.blockchains.cardano.network.common.models.CardanoUnspentOutput
 import com.tangem.blockchain.common.Amount
 import com.tangem.blockchain.common.AmountType
 import com.tangem.blockchain.common.TransactionData
@@ -28,6 +28,7 @@ class CardanoTransactionTest {
                 -31, -14, -112, 17, 29, -95, 14, -61, -25, -29, 87, 42, -121, -27, -109, -125, -56, -25, 63, -111, -86,
                 -60, 48, 90, -16, 1, 13, -126, 117, -23, -3,
             ),
+            assets = emptyList(),
         )
 
         val unspentOutput2 = CardanoUnspentOutput(
@@ -38,6 +39,7 @@ class CardanoTransactionTest {
                 31, -21, -70, -62, 13, -124, -92, -81, -23, -58, 114, 70, -101, -66, 43, 1, -24, 84, -25, 48, -74, 39,
                 -109, 53, -68, -9, -75, -49, -57, -89, 33, 53,
             ),
+            assets = emptyList(),
         )
 
         transactionBuilder.update(
