+++ blockchain/src/main/java/com/tangem/blockchain/blockchains/cardano/network/adalite/converters/AdaliteUnspentOutputConverter.kt
@@ -1,28 +1,28 @@
 package com.tangem.blockchain.blockchains.cardano.network.adalite.converters
 
-import com.tangem.blockchain.blockchains.cardano.network.CardanoUnspentOutput
-import com.tangem.blockchain.blockchains.cardano.network.adalite.repsonse.AdaliteBalance
 import com.tangem.blockchain.blockchains.cardano.network.adalite.repsonse.AdaliteUnspentOutputsResponse
+import com.tangem.blockchain.blockchains.cardano.network.common.models.CardanoUnspentOutput
 import com.tangem.common.extensions.hexToBytes
+import com.tangem.blockchain.blockchains.cardano.network.adalite.repsonse.AdaliteUnspentOutputsResponse.Utxo.Amount.Token as AdaliteUtxoToken
 
 internal object AdaliteUnspentOutputConverter {
 
     fun convert(response: AdaliteUnspentOutputsResponse): List<CardanoUnspentOutput> {
         return response.successData
             // we need to ignore unspent outputs with tokens (until we start supporting tokens)
-            .filter { it.amountData.tokens.isEmpty() }
+            .filter { it.amount.tokens.isEmpty() }
             .map { utxo ->
                 CardanoUnspentOutput(
                     address = utxo.address,
-                    amount = utxo.amountData.amount,
+                    amount = utxo.amount.value,
                     outputIndex = utxo.outputIndex.toLong(),
                     transactionHash = utxo.hash.hexToBytes(),
-                    assets = utxo.amountData.tokens.mapToAsset(),
+                    assets = utxo.amount.tokens.mapToAsset(),
                 )
             }
     }
 
-    private fun List<AdaliteBalance.Token>.mapToAsset(): List<CardanoUnspentOutput.Asset> {
+    private fun List<AdaliteUtxoToken>.mapToAsset(): List<CardanoUnspentOutput.Asset> {
         return map { token ->
             CardanoUnspentOutput.Asset(
                 policyID = token.policyId,
