+++ blockchain/src/main/java/com/tangem/blockchain/blockchains/cardano/network/common/models/CardanoUnspentOutput.kt
@@ -1,0 +1,16 @@
+package com.tangem.blockchain.blockchains.cardano.network.common.models
+
+internal class CardanoUnspentOutput(
+    val address: String,
+    val amount: Long,
+    val outputIndex: Long,
+    val transactionHash: ByteArray,
+    val assets: List<Asset>,
+) {
+
+    data class Asset(
+        val policyID: String,
+        val assetNameHex: String,
+        val amount: Long,
+    )
+}
